pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ambulance-openenv"
version = "1.0.0"
description = "Ambulance Dispatch Reinforcement Learning Environment (OpenEnv-compatible)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["reinforcement-learning", "emergency-dispatch", "openenv", "simulation"]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "networkx>=3.2",
    "numpy>=1.26",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "black",
    "ruff",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "agents*", "tasks*", "server*", "rl*", "utils*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
